Initial Changes
Automatically display all projects as cards in the projects_list.html. Right now, project markdown files can be placed anywhere (but still kept in projects) as long as the isProject is true in the front matter. Also removed the meme project that was used to test formatting of the cards. Changed the description of the website. Commented out the related posts and comments section of all pages.